Implement base _app.tsx
