+++ package.json
@@ -39,6 +39,7 @@
     "react": "^18.2.0",
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^18.2.0",
+    "sass": "^1.58.0",
     "three": "^0.140.2",
     "typescript": "4.9.5",
     "uuid": "^8.2.0"
